feat/water-3398 -- code refactored
