+++ src/internal/modules/charge-information/forms/charge-category/volume.js
@@ -8,10 +8,10 @@ const { getChargeCategoryData, getChargeCategoryActionUrl } = require('../../lib
 const getErrors = key => {
   const errors = {
     'string.pattern.base': {
-      message: 'Enter a number for the quantity using 6 decimal places or fewer, the number must be more than 0'
+      message: 'Enter a number for the volume using 6 decimal places or fewer, the number must be more than 0'
     },
     'any.required': {
-      message: 'Enter a quantity'
+      message: 'Enter a volume'
     },
     'string.empty': {
       message: 'Enter a qauntity'
@@ -30,14 +30,14 @@ const getErrors = key => {
 const form = request => {
   const { csrfToken } = request.view;
   const data = getChargeCategoryData(request);
-  const action = getChargeCategoryActionUrl(request, CHARGE_CATEGORY_STEPS.quantity);
+  const action = getChargeCategoryActionUrl(request, CHARGE_CATEGORY_STEPS.volume);
 
   const f = formFactory(action, 'POST');
-  f.fields.push(fields.text('quantity', {
+  f.fields.push(fields.text('volume', {
     controlClass: 'govuk-input govuk-input--width-10',
     suffix: 'ML',
     errors: getErrors()
-  }, data.quantity || ''));
+  }, data.volume || ''));
   f.fields.push(fields.hidden('csrf_token', {}, csrfToken));
   f.fields.push(fields.button(null, { label: 'Continue' }));
 
@@ -48,7 +48,7 @@ const schema = () => {
   const nonZeroNumberWithWSixDpRegex = new RegExp(/^\s*(?=.*[1-9])\d*(?:\.\d{1,6})?\s*$/);
   return Joi.object().keys({
     csrf_token: Joi.string().uuid().required(),
-    quantity: Joi.string().pattern(nonZeroNumberWithWSixDpRegex).required()
+    volume: Joi.string().pattern(nonZeroNumberWithWSixDpRegex).required()
   });
 };
 exports.schema = schema;
