+++ src/internal/views/nunjucks/charge-information/view.njk
@@ -8,8 +8,8 @@
 {% from "badge.njk" import badge %}
 {% from "forms/index.njk" import formRender,
 formErrorSummary %}
-{% from "nunjucks/charge-information/macros/charge-element.njk" import chargeElementTable %}
-{% from "nunjucks/charge-information/macros/charge-category.njk" import chargeCategoryTable %}
+{% from "nunjucks/charge-information/macros/charge-element-table.njk" import chargeElementTable %}
+{% from "nunjucks/charge-information/macros/charge-element-table-sroc.njk" import chargeCategoryTable %}
 
 {% macro summaryListRow(key, value, changeLink) %}
   <div class="govuk-summary-list__row">
@@ -148,13 +148,13 @@ formErrorSummary %}
 
   <form method="post"  action="{{action}}" novalidate="novalidate">
     <input type="hidden" name="csrf_token" value="{{ csrfToken }}"/>
-    {# {% if chargeVersion.chargeElements.length > 0 %} #}
-        {{ chargeElementTable(chargeVersion, isEditable) }}
-    {# {% else %}
+    {% if chargeVersion.chargeElements.length > 0 %}
+        {{ chargeElementTable(chargeVersion.chargeElements, isEditable, licenceId) }}
+    {% else %}
       {% for chargeCategory in chargeVersion.chargeCategories %}
-        {{ chargeCategoryTable(chargeCategory, isEditable)}}
+        {{ chargeCategoryTable(chargeCategory, isEditable, licenceId)}}
       {% endfor %}   
-    {% endif %} #}
+    {% endif %}
 </form>
